Added the files for v1
v1: This app displays a list of color coded unresolved NetCloud online/offline events, constantly updating to check for new events. When an event is deleted in the app a command is sent to delete it from the master SQL table, and when a new event appears a sound is played signaling whether it is an online or an offline event.